requirements.txt
branca==0.4.2
colour==0.1.5
Django==3.2.8
django_filter==21.1
folium==0.12.1
numpy==1.21.2
pandas==1.3.4
